Create separate file for menu
